pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "The comprehensive project for School Management System"
authors = ["Caleb Nwaizu <calebchris000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
pydantic = "^2.4.1"
aerich = "^0.7.2"
uvicorn = "^0.23.2"
pyjwt = "^2.8.0"
bcrypt = "^4.0.1"
python-dotenv = "^1.0.0"
flask-login = "^0.6.2"
asyncpg = "^0.28.0"
tortoise-orm = "^0.20.0"


[tool.poetry.scripts]
server = {callable = 'bootstrap:run_server'}


[tool.aerich]
tortoise_orm = "src.database.migration.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
